+++ CHANGELOG.rst
@@ -7,6 +7,17 @@ The format is based on `Keep a Changelog
 <https://keepachangelog.com/en/1.0.0/>`_, and this project adheres to `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+0.3.2
+-----
+
+Changed:
+
+ - Config files are generated instead of copied.
+
+Added:
+
+ - Kill script added to scripts
+
 0.3.1
 -----